requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-keyvault-secrets==4.9.0
azure-search-documents==11.5.2
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.9.0
cryptography==42.0.7
decorator==4.0.2
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
eventlet==0.36.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-redis==0.4.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fsspec==2023.12.2
future==0.18.3
gevent==24.2.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
isodate==0.7.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.4.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
numpy==1.24.4
openai==1.78.1
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipdeptree==2.16.1
platformdirs==4.2.0
portalocker==2.10.1
proglog==0.1.10
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyExecJS==1.5.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.26.0
pyparsing==3.1.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
redis==5.0.8
referencing==0.34.0
regex==2024.11.6
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
sympy==1.12
tenacity==9.0.0
tiktoken==0.9.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.13.2
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
website==1.1
Werkzeug==3.0.1
whisper==1.1.10
yarg==0.1.9
yarl==1.20.0
zope.event==5.0
zope.interface==6.4.post2